+++ src/main/kotlin/dev/hagios/jetpackcomposepreviewcreator/settings/PreviewSettingsState.kt
@@ -6,6 +6,7 @@ class PreviewSettingsState : BaseState() {
     var functionNameExtension by string("Preview")
     var defaultVisibility by enum(Visibility.private)
     var generatePosition by enum(Position.`end of file`)
+    var overrideBehaviour by enum(Behaviour.replace)
     var addParameterNames by property(true)
     var useDefaultValues by property(true)
     var useNullValues by property(true)